+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG
 Latest version is always on top. 
 
+## v0.3a:
+- Added support for Image compression with custom output options (Unstable)
+- Added support for compression strength
+
 ## v0.2a:
 - Fixed a bug where we would try to write a directory path to a file when extracting folders from a zip-file (Github Issue #1)
 - Added initial support for image compression